pbdZMQ-package {pbdZMQ}	R Documentation

Programming with Big Data – Interface to ZeroMQ

Description

ZeroMQ is a well-known library for high-performance asynchronous messaging in scalable, distributed applications. This package provides high level R wrapper functions to easily utilize ZeroMQ. We mainly focus on interactive client/server programming frameworks. For convenience, a minimal ZeroMQ library (4.1.0 rc1) is shipped with pbdZMQ, which can be used if no system installation of ZeroMQ is available. A few wrapper functions compatible with rzmq are also provided.

Details

Package: 	 pbdZMQ

Type: 	 Package

License: 	 GPL-3 2.0

LazyLoad: 	 yes

The install command using default pbdZMQ's internal ZeroMQ library is

> R CMD INSTALL pbdZMQ_0.1-0.tar.gz
--configure-args="--enable-internal-zmq"

Other available variables include

Variable 	 Default

ZMQ_INCLUDE 	 -I./zmqsrc/include

ZMQ_LDFLAGS 	 -L./ -lzmq

ZMQ_POLLER 	 select

See the package source file pbdZMQ/configure.ac for details.

For installation using an external ZeroMQ library, see the package source file pbdZMQ/INSTALL for details.

Author(s)

Wei-Chen Chen wccsnow@gmail.com.

References

ZeroMQ/4.1.0 API Reference: http://api.zeromq.org/4-1:_start

Programming with Big Data in R Website: http://r-pbd.org/

See Also

zmq.ctx.new(), zmq.socket().

[Package pbdZMQ version 0.2-2 Index]