ZMQ Control Environment {pbdZMQ}	R Documentation

Sets of controls in pbdZMQ.

Description

These sets of controls are used to provide default values in this package.

Format

Objects contain several parameters for communicators and methods.

Details

The elements of .pbd_env$ZMQ.ST are default values for socket types as defined in ‘zmq.h’ including

Elements 	 Value 	 Usage

PAIR 	 0L 	 socket type PAIR

PUB 	 1L 	 socket type PUB

SUB 	 2L 	 socket type SUB

REQ 	 3L 	 socket type REQ

REP 	 4L 	 socket type REP

DEALER 	 5L 	 socket type DEALER

ROUTER 	 6L 	 socket type ROUTER

PULL 	 7L 	 socket type PULL

PUSH 	 8L 	 socket type PUSH

XPUB 	 9L 	 socket type XPUB

XSUB 	 10L 	 socket type XSUB

STREAM 	 11L 	 socket type STREAM

The elements of .pbd_env$ZMQ.SO are default values for socket options as defined in ‘zmq.h’ including 60 different values, see .pbd_env$ZMQ.SO and ‘zmq.h’ for details.

The elements of .pbd_env$ZMQ.SR are default values for send/recv options as defined in ‘zmq.h’ including

Elements 	 Value 	 Usage

BLOCK 	 0L 	 send/recv option BLOCK

DONTWAIT 	 1L 	 send/recv option DONTWAIT

NOBLOCK 	 1L 	 send/recv option NOBLOCK

SNDMORE 	 2L 	 send/recv option SNDMORE (not supported)

The elements of .pbd_env$ZMQ.MC are default values for warning and stop controls in R. These are not the ZeroMQ's internal default values. They are defined as

Elements 	 Value 	 Usage

warning.at.error 	 TRUE 	 if warn at error

stop.at.error 	 TRUE 	 if stop at error

Author(s)

Wei-Chen Chen wccsnow@gmail.com.

References

ZeroMQ/4.1.0 API Reference: http://api.zeromq.org/4-1:_start

Programming with Big Data in R Website: http://r-pbd.org/

See Also

.zmqopt_init().

[Package pbdZMQ version 0.2-2 Index]