Initial Control Functions {pbdZMQ}	R Documentation

Initial controls in pbdZMQ

Description

Initial control functions

Usage

.zmqopt_get(main, sub = NULL, envir = .GlobalEnv)

.zmqopt_set(val, main, sub = NULL, envir = .GlobalEnv)

.zmqopt_init(envir = .GlobalEnv)

Arguments

main
a variable to be get from or set to

sub
a subvariable to be get from or set to

envir
an environment where ZMQ controls locate

val
a value to be set

Details

.zmqopt_init() initials default ZMQ controls. .zmqopt_get() gets a ZMQ control. .zmqopt_set() sets a ZMQ control.

Value

.zmqopt_init() initial the ZMQ control at envir.

Author(s)

Wei-Chen Chen wccsnow@gmail.com.

References

ZeroMQ/4.1.0 API Reference: http://api.zeromq.org/4-1:_start

Programming with Big Data in R Website: http://r-pbd.org/

See Also

.pbd_env.

Examples

## Not run: 
library(pbdZMQ, quietly = TRUE)

ls(.pbd_env)
rm(.pbd_env)
.zmqopt_init()
ls(.pbd_env)

.pbd_env$ZMQ.SR$BLOCK
pbd_opt(bytext = "ZMQ.SR$BLOCK = 0L")

## End(Not run)

[Package pbdZMQ version 0.2-2 Index]